+++ crates/pvm/parser/src/format/o.rs
@@ -4,14 +4,19 @@ use crate::format::{I, O};
 
 impl From<O> for Vec<u8> {
     fn from(value: O) -> Self {
-        I { imm0: value.off0 }.into()
+        I {
+            imm0: value.off0 as u64,
+        }
+        .into()
     }
 }
 
 impl From<&[u8]> for O {
     fn from(bytes: &[u8]) -> Self {
         let i = I::from(bytes);
-        O { off0: i.imm0 }
+        O {
+            off0: i.imm0 as i32,
+        }
     }
 }
